Add files via upload
Hamp Crafts is a family-owned craft store that has been in business for decades. Its current operations are limited to its physical (brick-and-mortar) storefront. Recently, the owners have determined that one of the best ways to gain additional revenue is through online sales. The online sales process is new to the owners. They need to fully understand how online ordering works, and how it can be integrated with their current in-store purchase and supply process.

You are a systems analyst working with Hamp Crafts to modernize their business. First, you will analyze their current purchase and supply process for items in their brick-and-mortar store. Then you will suggest ways to integrate new processes for an online storefront.