Add files via upload
Assume the role of a systems analyst designing an LMS for YOUser University. You have been asked to determine different functional and nonfunctional requirements for the new LMS. You have also been asked to determine limitations and assumptions for your system design. You will be asked to provide a rationale for your work, and to support your reasoning with quality sources.